examples/consul_app/requirements.txt
python-consul>=0.6.1
vyper
